signal-engine/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.7.3
aiohttp==3.9.5
websockets==12.0
